Move `Effect` to its own file
