+++ KrakenVault/Composable Architecture/Store.swift
@@ -7,20 +7,6 @@ import SwiftUI
 
 public typealias Reducer<Value, Action> = (inout Value, Action) -> [Effect<Action>]
 
-public struct Effect<A> {
-    public let run: (@escaping (A) -> Void) -> Void
-
-    public init(run: @escaping (@escaping (A) -> Void) -> Void) {
-        self.run = run
-    }
-
-    public func map<B>(_ f: @escaping (A) -> B) -> Effect<B> {
-        Effect<B> { callback in
-            self.run { a in callback(f(a)) }
-        }
-    }
-}
-
 public final class Store<Value, Action>: ObservableObject {
     private let reducer: Reducer<Value, Action>
     @Published public private(set) var value: Value
